test: configure project for tests
