+++ src/lib.rs
@@ -1,0 +1,7 @@
+pub mod modules;
+pub use modules::audio_control::AudioModule;
+pub use modules::core::{
+    create_error_response, create_response, get_timestamp, handle_message,
+    models::{ModuleType, ServerEvent},
+    Broadcaster, ModuleRegistry,
+};
