+++ src/modules.rs
@@ -1,2 +1,3 @@
+pub mod app_router;
 pub mod audio_control;
 pub mod core;
