+++ src/modules/audio_control/mod.rs
@@ -8,3 +8,6 @@ pub mod utils;
 
 pub use audio_handlers::handle_action_sound_request;
 pub use audio_module::AudioModule;
+
+#[cfg(test)]
+mod tests;
