+++ src/modules/core/utils/mod.rs
@@ -1,3 +1,2 @@
 pub mod timestamp_utils;
-
 pub use timestamp_utils::get_timestamp;
